Refactor: token refresher in a thread. Switched the logic to AudioRemote. They main file will only be the start point.
